[modify] divide AudioUnit initialize proc
divide AudioUnit initilize proc
